requirements.txt
dj-database-url==0.3.0
Django==1.8.5
django-cors-headers==1.1.0
djangorestframework==3.3.0
gunicorn==19.3.0
psycopg2==2.6.1
wheel==0.24.0
whitenoise==2.0.4
